Update README.md
Election Analysis 
#Overview of Election Audit: Explain the purpose of this election audit analysis.
The purpose of the election audit was to support the election commission with additional data to complete the election audit. Per request the election commission would like to know the voter turnout for each county, the percentage of votes from each county out of the total count and the county with the highest turnout. 

#Election-Audit Results: Using a bulleted list, address the following election outcomes. Use images or examples of your code as support where necessary.

*How many votes were cast in this congressional election?
According the data, The total number of casted votes in the congressional election was 369,711 

*Provide a breakdown of the number of votes and the percentage of total votes for each county in the precinct.
According the data the county of Denver had the largest number of votes with a total of 306,055(82.8%). Jefferson county came in second place with 38,855(10.5%). Finally, Arapahoe County had the least amount of votes with 24,801(6.7%)

*Which county had the largest number of votes?
According the data Denver County had the largest number of votes with a total of 306,055(82.8%). 

*Provide a breakdown of the number of votes and the percentage of the total votes each candidate received.
According to the data, Diana DeGette led all candidates with 73.8% (272,892) of all votes. In second place was Charles Casper Stockham with 23.0% (85,213) of all votes. Finally, in last place was Raymond Anthony Doane who received just 3.1% (11,606) of all votes.  

*Which candidate won the election, what was their vote count, and what was their percentage of the total votes?
According to the data the winner of this election was Diana DeGette. Her vote count was 272,892 which is about 74% of all county votes. 

#Election-Audit Summary: In a summary statement, provide a business proposal to the election commission on how this script can be used—with some modifications—for any election. Give at least two examples of how this script can be modified to be used for other elections.
I strongly advise that the election commission save this script to be utilized for futuer elections. The script has accurately calculated the voter turnout, percentage of votes for each county, the county with the highest turnout and, of course, it was able determine the winning candidate of the congressional election. If need be, modifications can be made to the script. 
For example, if the commission wishes to add a particular county to the county listed it could use the append function, i.e. >>>>>>counties.append('El Paso"). On the hand if it decided to remove a county it could us the remove i.e. >>>>counties.remove("Denver"). All in all saving this script will allow for efficient calculations of ballots submitted, voter turnout, and county data.